new simil function
optimized for sparse matrix use

+++ mesh.cpp
@@ -189,13 +189,13 @@ void Mesh::computeCoRs() {
 
 QVector3D Mesh::computeCoR(uint i) {
     QVector3D c;
-    double s = 0.0;
+    float s = 0.0;
     uint count = indices.size()/3;
     for(uint j = 0; j < count; j++) {
         Triangle t = {indices[3*j], indices[3*j+1], indices[3*j+2]};
-        double similArea = (double) skeleton.simil(i, t)*area(t);
+        float similArea = skeleton.simil(i, t)*area(t);
         
-        c += (vertices[t.a]+vertices[t.b]+vertices[t.c])/3.*((float) similArea);
+        c += (vertices[t.a]+vertices[t.b]+vertices[t.c])/3.*similArea;
         s += similArea;
         if(isnan(similArea) || isinf(similArea)) {
             cout << "simil " << skeleton.simil(i, t) << endl;
